Entered actual code
So I keep getting the error that my constructor isn't usable because the entered variable are not static but the constructor is. However I never made my constructor static. I'm more than happy to quickly review my code once the answer key is up, but I'm sure my code is close to identical.